Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -6,6 +6,7 @@ use App\Models\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
 use DB;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -66,6 +67,9 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort('403', 'Access denied.');
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -78,6 +82,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort('403', 'Access denied.');
+        }
         $question->update($request->only('title', 'body'));
         // return redirect()->route('questions.index')->with('success', 'Your question has been submitted.');
         return redirect("/questions")->with('success', 'Your question have been updated.');
@@ -91,6 +98,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort('403', 'Access denied.');
+        }
         $question->delete();
         return redirect("/questions")->with("success", "Your question has been deleted.");
     }
